fix new thread tracking
Emails always should end up in the folder that their parents end up in.
However, when a new email thread arrives, all emails in that thread
should be considered for filtering.

The `parent_is_new` variable was responsible for doing the switch
between the two. However, it fails to capture the scenario in which two
new emails are sent in reply to an existing email. If that email sits in
a different folder than the newest one would be sent to, an error
occurred.

To fix this, a new hash set is introduced tracking new threads
correctly.